added multiple functions
sorted functions to different header files
